Revert "сокращение класса FilterFragment - отдельный класс FilterUiHelper"
