+++ app/src/main/java/ru/practicum/android/diploma/app/App.kt
@@ -6,7 +6,6 @@ import org.koin.core.context.startKoin
 import ru.practicum.android.diploma.di.dataModule
 import ru.practicum.android.diploma.di.interactorModule
 import ru.practicum.android.diploma.di.repositoryModule
-import ru.practicum.android.diploma.di.uiModule
 import ru.practicum.android.diploma.di.viewModelModule
 
 class App : Application() {
@@ -15,7 +14,7 @@ class App : Application() {
         super.onCreate()
         startKoin {
             androidContext(this@App)
-            modules(repositoryModule, interactorModule, viewModelModule, dataModule, uiModule)
+            modules(repositoryModule, interactorModule, viewModelModule, dataModule)
         }
     }
