--- app/src/main/java/ru/practicum/android/diploma/di/UiModule.kt
@@ -1,12 +1,0 @@
-package ru.practicum.android.diploma.di
-
-import org.koin.dsl.module
-import ru.practicum.android.diploma.databinding.FragmentFilterBinding
-import ru.practicum.android.diploma.filter.presentation.ui.helper.FilterUiHelper
-import ru.practicum.android.diploma.filter.presentation.viewmodel.FilterViewModel
-
-val uiModule = module {
-    factory { (binding: FragmentFilterBinding, viewModel: FilterViewModel) ->
-        FilterUiHelper(binding, viewModel, get())
-    }
-}
